Add files via upload
Kuudrafollowerquest: Introducing "Kuudrafollowerquest," a captivating Java mod designed to enhance the Kuudra quest experience in Hypixel Skyblock. This mod seamlessly integrates a challenging yet rewarding puzzle-solving mechanism, adding a layer of help to the quest.

This Mod Is An Open Source Code Mod

KuudraHelper helps with Kuudra's quest and it shows all the locations of the npcs in Crimson Isle & Spider's Den

So that you know, this is a forge mod. Not a spigot plugin. Use the forge mod loader to use this mod for yourself.

Usage Download the jar file from the top of the page and place it into your Minecraft installation's mod folder (usually %appdata%/.minecraft/mods) Start Minecraft with forge and enjoy!

How to use You can configure the mod in the mods menu. Check your forge installation on where to find the settings panel

Commands /kf - Opens the mod menu with a guide on how to use the mod effectively